+++ src/components/SearchAndFilter/index.js
@@ -1,4 +1,4 @@
 export { default as PublicationRequestsFilters } from './PublicationRequestsFilters';
 export { default as OAFilterHeaderComponent } from './OAFilterHeaderComponent';
-
-
+export { default as JournalsFilters } from './JournalsFilters';
+export { default as PartiesFilters } from './PartiesFilters';