+++ src/routes/PartiesRoute/PartiesRoute.js
@@ -12,7 +12,10 @@ import {
 } from '@folio/stripes/components';
 
 import { SASQRoute } from '@k-int/stripes-kint-components';
-import { OAFilterHeaderComponent } from '../../components/SearchAndFilter';
+import {
+  OAFilterHeaderComponent,
+  PartiesFilters,
+} from '../../components/SearchAndFilter';
 import Party from '../../components/views/Party';
 import urls from '../../util/urls';
 import focusSASQSearchField from '../../util/focusSASQSearchField';
@@ -28,7 +31,7 @@ const PartiesRoute = ({ path }) => {
     endpoint: PARTIES_ENDPOINT,
     SASQ_MAP: {
       searchKey: 'mainEmail,givenNames,familyName,orcidId',
-      filterKeys: {},
+      filterKeys: { institutionLevel1: 'institutionLevel1.value' },
     },
   };
 
@@ -96,8 +99,12 @@ const PartiesRoute = ({ path }) => {
     >
       <SASQRoute
         fetchParameters={fetchParameters}
+        FilterComponent={PartiesFilters}
         FilterPaneHeaderComponent={renderHeaderComponent}
         id="parties"
+        labelOverrides={{
+          foundValues: 'ui-oa.parties.found#People',
+        }}
         mainPaneProps={{
           appIcon: <AppIcon app="oa" iconKey="party" size="small" />,
           lastMenu: lastpaneMenu,
@@ -107,6 +114,7 @@ const PartiesRoute = ({ path }) => {
         path={path}
         resultColumns={resultColumns}
         sasqProps={{ initialSortState: { sort: 'familyName,givenNames' } }}
+        searchFieldAriaLabel="parties-search-field"
         ViewComponent={Party}
       />
     </HasCommand>