+++ src/settings/index.js
@@ -3,8 +3,6 @@ import PropTypes from 'prop-types';
 
 import { useIntl } from 'react-intl';
 
-import { useStripes } from '@folio/stripes/core';
-
 import { useSettings } from '@k-int/stripes-kint-components';
 
 import { ChecklistItemDefinitions, PickListValues } from './settingsComponents';
@@ -13,8 +11,8 @@ import { REFDATA_ENDPOINT, SETTINGS_ENDPOINT } from '../constants/endpoints';
 const propTypes = {
   resources: PropTypes.shape({
     settings: PropTypes.shape({
-      records: PropTypes.arrayOf(PropTypes.object)
-    })
+      records: PropTypes.arrayOf(PropTypes.object),
+    }),
   }),
   match: PropTypes.shape({
     path: PropTypes.string.isRequired,
@@ -22,21 +20,20 @@ const propTypes = {
 };
 
 const OASettings = (props) => {
-  const stripes = useStripes();
   const intl = useIntl();
   const persistentPages = [];
 
-  if (stripes.hasPerm('settings.oa.picklists.manage')) {
-    persistentPages.push({
-      component: PickListValues,
-      label: intl.formatMessage({ id: 'ui-oa.settings.refdata.picklistValues' }),
-      route: 'pick-list-values',
-    });
-  }
+  persistentPages.push({
+    component: PickListValues,
+    label: intl.formatMessage({ id: 'ui-oa.settings.refdata.picklistValues' }),
+    route: 'pick-list-values',
+  });
 
   persistentPages.push({
     component: ChecklistItemDefinitions,
-    label: intl.formatMessage({ id: 'ui-oa.settings.checklistItemDefinitions' }),
+    label: intl.formatMessage({
+      id: 'ui-oa.settings.checklistItemDefinitions',
+    }),
     route: 'checklist-items',
   });
 
@@ -45,18 +42,14 @@ const OASettings = (props) => {
     intlKey: 'ui-oa',
     persistentPages,
     refdataEndpoint: REFDATA_ENDPOINT,
-    settingEndpoint: SETTINGS_ENDPOINT
+    settingEndpoint: SETTINGS_ENDPOINT,
   });
 
   if (isLoading) {
     return null;
   }
 
-  return (
-    <SettingsComponent
-      {...props}
-    />
-  );
+  return <SettingsComponent {...props} />;
 };
 
 OASettings.propTypes = propTypes;